[MOD] Renamed GenericValue attribute to DrawAsType for clarity. Moved it from Odin only to non odin (WIP)
